Use intl for text direction (see #3360)
Description
-----------

Currently the text-direction information for a language is stored in the Contao translations themselves. This is a problem however if the used front end language is not an installed language - or when a translation gets removed from Contao because it is not sufficiently translated anymore. For example in Contao 4.12 Farsi got removed (again) and then the front end will suddenly display the text left-to-right instead of right-to-left. See https://github.com/contao/core-bundle/commit/4073033679391d66c67aa08b9892da6bd8b70c00#commitcomment-55223764 for instance.

PHP intl already provides this information so I think we should not store this information in our own translations and rather extract it from intl.

This PR uses ~a lightweight library~ PHP intl to extract the direction for the current language. It also removes the `MSC.textDirection` translation "label" and adds it again in `System::loadLanguageFile` for BC (in case anyone else still accesses `MSC.textDirection`).

As I consider this a bugfix, I have implemented this for Contao 4.9.

/cc @hamidabbaszadeh

Commits
-------

391e6b55 use intl for text direction
37df6bcf update comments
4043fb64 use MSC.textDirection and then intl as fallback
4a545474 do not override MSC.textDirection
1cb5bab5 Apply suggestions from code review

Co-authored-by: Martin Auswöger <martin@auswoeger.com>
407ddbbc remove cosmo library
3e0c07cb add deprecations
1b61a35c code style
fa238990 update translator deprecation
748ffdde do not trigger deprecation when already cached
25d39400 adjust deprecation implementation
3de1245b remove remaining MSC.textDirection translations
b61d444d Revert "remove remaining MSC.textDirection translations"

This reverts commit 3de1245b9a00fb4864a3ed82182472125954174e.
35d2523a remove BC for overriding MSC.textDirection
078884d0 no fallback to MSC.textDirection needed anymore
79d8e3ee remove deprecation notice from translator
c48467ca use locale