Allow extending both front end and back end templates with Twig (see #3362)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #3361 
| Docs PR or issue | -

This PR extracts the methods in `FrontendTemplate` and `BackendTemplate` that are supposed to be used from with the templates themselves into traits. This way both can be used when rendering via the Twig template proxy.

Commits
-------

d4699d78 move methods that should be accessible by templates into traits
e01266bc render legacy template with FE + BE template capabilities
8e99e858 add tests
9c6dc458 restore formatting
1514c3d8 Restore the original formatting