Also pass the template object in the "parseFrontendTemplate" hook (see #3352)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes https://github.com/contao/contao/issues/3184 
| Docs PR or issue | contao/docs#...

see https://github.com/contao/contao/issues/3184 
Passing the templateObject itself to the parseFrontendTemplate Hook method.

Maybe it´s possible to get it also at the 4.9 branch. That´s because i used the Reflections! But if it´s a feature for future version  we can leave the Reflection-Calls and just add the object itself as third parameter.

Thanks

Commits
-------

375a345f adding slef object to parseFrontendTemplate