Priority 0 is not the same as not-set priority (see #3444)
Description
-----------

Preparing for https://github.com/contao/contao/pull/3439#discussion_r704441417

A priority of `0` is not the same as not having a priority. The `PriorityTaggedServiceTrait` will call a `getDefaultPriority` method if it exists on the service and a priority is not defined, which should not be called if priority is `0`.

FYI we cannot add `int|null` to the public property because Doctrine Annotations cannot handle that (see https://github.com/contao/contao/pull/764). We do not set a default property value because Psalm might complain about the annotation not including `null`. This becomes obsolete with #3439 in Contao 4.13 anyway.

Commits
-------

fc008668 Priority 0 is not the same as not-set priority