Fix the picker for child tables without PID (see #3457)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2046

A picker that picks elements from a child table like the one used in the include content element currently creates links like `?do=article&table=tl_content` which show an empty page because the pid (`id=`) is missing.

If there is a `CURRENT_ID` in the session, an arbitrary page might show up instead of an empty page, or it causes a permission error as in #2046.

This pull request fixes this by only adding the `table=` parameter if there is a valid pid.

Commits
-------

651d8dfc Fix picker for child tables without pid
f333be6f Move the default declaration above the if condition
53489264 Handle the case that $ptable is false

Co-authored-by: Martin Auswöger <martin@auswoeger.com>