Handle edge cases in StringUtil::convertEncoding() (see #3496)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #3493


There is an additional check via `is_string` and `is_numeric` to only allow encoding for simple stringable parameters and to prevent errors from the `mb_` functions when called with e.g. an array. For invalid parameters, an empty string is returned.

This also makes sure the method returns an empty string when called with a non-string parameter and the same `$to` and `$from` encoding. Until now, the method has returned the unchanged parameter (which could be an array, object etc.).

Additionally, the return value is always cast to a string if the encodings match to make sure integers and floats are returned as strings (which they were for different input & output encodings but not for matching encodings).

**Test case**
This also adds a simple unit test for `StringUtil::convertEncoding`.

One could add a lot more test cases for the concrete encodings, but since the `StringUtil` only uses `mb_encode`, `mb_decode`, `mb_detect_encoding` and `mb_convert_encoding` internally, these tests should not be necessary.

Commits
-------

fb18cd86 handle edge cases in StringUtil::convertEncoding
745c9ebf handle stringables, add deprecation
9d1be413 privatise utility methods
76ac28f5 inline stringable check & cast
63f72179 remove leftover import