Add Chosen to select menus in the backend DCA filters (see #3660)
Description
-----------

This pull request adds the Chosen plugin to select menus (to both, filter dropdowns and the search dropdown) in the backend DCA filters. Screenshot:

<img width="729" alt="Zrzut ekranu 2021-11-10 o 11 03 43" src="https://user-images.githubusercontent.com/193483/141093334-35203ae1-db5b-42af-b435-3af1712480b3.png">

One thing to consider is that it is possible to hit the "Enter" key and submit the form with the current regular select menus. This is not yet the case with the "Chosen" plugin, but we might want to add that.

Feature proposal per @Toflar request.

Commits
-------

8ce0ae64 Add Chosen to select menus in the backend DCA filters
b4ab9268 Merge branch '4.x' into feature/dca-filters-chosen
867c8817 Add Chosen to select menus to sorting and limit dropdowns
15a8c412 Add Chosen to search menu in DC_Folder