Do not initialize the framework to get the Automator commands (see #3882)
Description
-----------

Since we no longer need the legacy `ClassLoader`, we don't need to initialize the Contao framework to read the class methods. Initializing the framework can lead to an exception when trying to list the available console commands, e.g. if a database connection is not correctly set up.

Commits
-------

1133fbf6 Do not initialize framework to get Automator command