Fixed tl_page permissions for routing fields (see #3916)
Description
-----------

I forgot to add `exclude => true` to the new routing fields, they would always be available to all users.

The label change is necessary for the field name to correctly show in permissions, otherwise only the first character of the translation would be shown.

Regarding `type` and `requireItem` fields, they are only reordered for the permission to closely match to the field order in the default palettes.

Commits
-------

76795ccf Fixed tl_page permissions
d72f3aad Use array label for route conflicts