Correctly handle types and empty values in DC_Table::save() (see #3909)
Description
-----------

Fixes https://github.com/contao/contao/pull/3904

I was unable to figure out why this did work in Doctrine 2, but Doctrine 3 can only convert parameters based on their types, and we're not giving it the type information, even though it's there. So I extended our database statement and added the type to DC_Table if available.

Also noticed that the fallback empty value seems incorrect to me and should be inferred from the type/SQL as well.

FYI I also tried to apply this to Contao 4.9 but our database abstraction has changed significantly (rewrite on DBAL by @ausi) and it would require a lot more changes in 4.9 that I would be afraid to apply in a patch version.

Commits
-------

01a3c491 Correctly handle types in DC_Table::save()
602d55b7 Also add types in try-catch statement
cac1c988 Only build the types array once

Co-authored-by: Martin Auswöger <martin@auswoeger.com>