Allow Flysystem v3 (see #3935)
Description
-----------

This PR introduces the ability to check if directories exist in the filesystem implementation. This was a limitation with the not-that-old Flysystem 2, but Flysystem 3 was released shortly because this was considered a [design flaw](https://blog.frankdejonge.nl/flysystem-3-0-0-is-released/). I probably should have went with my gut feeling and added these things no matter what in #3774, but then trusted them on this decision. :smile: 

Ok, not too late. Here are the changes:

* The `VirtualFilesystemInterface`, `DbafsManager` and `MountManager` now allow to differentiate between `fileExists()` and `directoryExists()`.
* The `VirtualFilesystemInterface` and `DbafsManager` also got a `has()` method as a shortcut to check if any resource exists. It's a convenient method, so it would only be 'necessary' for the virtual filesystem, but it allows for a more effective evaluation in the DBAFS manager, so that's why it's there as well.
* I explicitly required `league/flysystem` in `^2.0 || ^3.0` because the Symfony bundle with the transient dependency was only released with a new minor version which completely hides the real dependency we're using.

Installing v3 is completely optional and the reason for that is, that it unfortunately [requires PHP8](https://github.com/thephpleague/flysystem/issues/1405). For users of v2 there is a workaround in place: we list contents of the parent directory of the requested resource and - if an item is returned where the path matches - check the type of the item. Not as effective, but fully compatible.

In Contao 5 we can then require v3, remove the fallback and also get rid of some code in the `Dbafs` class that currently also uses the workaround to determine if a resource is a directory.

Commits
-------

94810d59 allow flysystem in version 3
4e97853e support directoryExists in the mount manager
96e5d841 rename resourceExists() to has() in DbafsManager
955f1467 support to differentiate between files and directories in the dbafs manager
f2f53b83 support to differentiate between files and directories in the VirtualFilesystemInterface
7e6cd627 fix type hint
d5729bde revert an unrelated change
4e4b6d94 CS

Co-authored-by: Leo Feyer <github@contao.org>