Do not generate URL for insert tags that don’t need it (see #3915)
Description
-----------

It makes no sense to call routing and generate a URL if the insert tag does not need it. It also unnecessarily prevents these insert tags from working in Contao 4.12/4.13 if the page URL cannot be generated (non-routable / requires parameter).

FYI I intentionally did not remove the `case` for the same tags in the later method because it is required for non-regular pages (the new switch-case is inside an if-else handling).

Commits
-------

f8466e98 Early return for link_ inserttags without URL
9470c4b9 Use if condition
1f6cd9ef CS