Show custom Twig templates in the back end dropdowns (see #4121)
Description
-----------

Fixes #3274 

This is a rather simple solution to also get the custom Twig templates (`ce_text_foobar.html.twig`) to show up in the back end dropdowns. We need a more sophisticated one, that replaces the `TemplateLoader` completely, later on. But this will only be doable in Contao 5.

With this PR we're now merging the results from the new template hierarchy with the ones from the `TemplateLoader`. In theory the former would be sufficient because it also contains all PHP templates, but I think implementing it this way is safer, because it will still include files that were manually added.

Commits
-------

824d8d32 shows twig custom templates in the back end dropdowns
98d7109c use contao.twig.hierarchy as the service name
85c7b95e remove unnecessary quotes
1fd443ad Revert "remove unnecessary quotes"
0df4a7d5 Revert "use contao.twig.hierarchy as the service name"
600aedff use 'contao.twig.filesystem_loader' directly in the legacy code
c78f640b adjust tests