Do not process redirected URLs outside base domains (see #4218)
Description
-----------

Fixes #4213

See https://github.com/contao/contao/issues/4213#issuecomment-1052022561 for a detailed explanation of the problem.

This PR fixes this by comparing the host of the actual URL of the response with the base URI collection and returning a negative decision in `needsContent`, so that any internal URL that was redirected to an external URL is not getting processed.

Commits
-------

24669075 do not process redirected URLs outside base domains
9c034f1d reference issue
2a7072c4 report ok on redirected URLs
c1d25a53 use Uri class