Lower max file size in Dbafs service (see #4208)
Description
-----------

Fixes #4200 

This PR reduces the DBAFS max file size by one so that it also fits in a integer on 32-bit PHP versions. The value is opinionated anyways, so the effect is negligible. Without this change an error would occur because property initializers do no coerce or clamp values.

Commits
-------

52ccbb8b lower max file size
cac03acb update reference