Fix some dynamic routes handling (see #4197)
Description
-----------

Fixes https://github.com/contao/contao/issues/4191

Also found a few other issues by chance. Instead of injecting the router into the `ContaoFramework`, we could also throw an appropriate exception and handle it in the `PrettyErrorScreenHandler`. But that would disable the redirect `dev` mode, but we could also render a template to explain instead of a plain redirect.

Commits
-------

838c34a1 Fixed two cases for dynamic backend route prefix
ca673086 Updated tests