Clear the session value when toggle is closed (see #5811)
Description
-----------

Follow-up to https://github.com/contao/contao/pull/5797

if a user opens and closes 1000 page nodes in the tree, the session will keep a list of 1000 page IDs in the session. Since not having a session value equals to the child tree being closed, we shouldn't store the 0 as value.

Commits
-------

cf0765d8 Clear the session value when toggle is closed