Correctly handle sub-subpalettes in editAll mode (see #5863)
Description
-----------

Funny noone ever noticed that… Here's how to reproduce in the online demo:

1. edit the slider article of the home page
2. select the first two slider elements and edit multiple
3. check **all** fields in the tl_content field list

you should now get the edit form for two content elements. Both have an image assigned and inside the image subpalette, the "Overwrite metadata" subpalette should be enabled. This is the key to the problem. If a subpalette inside a subpalette is enabled.

Now unselect the "Add image" checkbox and save the record (or reload the page). After the page reload, enable the "Add image" checkbox again. Because the "Overwrite metadata" checkbox is already enabled, it will lead to sub-subpalette loading and incorrect handling (all editable field are rendered, not just the subpalettes).

The fix is a 99% copy of the same code we already have in the `edit()` method of `DC_Table`. We need to keep a "stack" of subpalettes and only return after all children are rendered.

Commits
-------

cbe2a848 Correctly handle sub-subpalettes in editAll mode
ba9b10ca Merge branch '4.13' into bugfix/editall-sub-subpalette

Co-authored-by: Leo Feyer <1192057+leofeyer@users.noreply.github.com>