Unwrap response exceptions thrown while rendering a template (see #5980)
Description
-----------

Tries to fix #5977 

Here is an attempt at fixing #5977. There is no extension point for the exception handling in Twig (it happens in the `Template` base class), but we can convert the exception in a kernel exception listener. It's not ideal but the effort to do so, is also very minimal.

If we agree on the concept - which exceptions should we "unwrap"?

/cc @qzminski

Commits
-------

6ba69524 unwrap certain exceptions if they were thrown while rendering a templ…
0ccd80af use instanceof check instead of testing against class name
779a8aa7 unwrap all response exceptions
49af556a Update core-bundle/src/EventListener/UnwrapTwigExceptionListener.php