Do not mark as copy more than once (see #6062)
Description
-----------

Imho there is not much value in having headlines and form field labels like

* Original
* Original (copy)
* Original (copy) (copy)
* Original (copy) (copy) (copy)

if you use "Save and duplicate" more than once or if you manually copy the copy etc.

This PR would make it so the suffix is only added once. This would alleviate the issue in #6058.

Commits
-------

4a16468d do not mark as copy more than once
bbfcaba2 use sprintf to remove placeholder