Fix root page dependent module when there is no module for a root page (see #6075)
Description
-----------

The selects of the `RootPageDependentSelect` are not mandatory and you do not have to define a module for each root page (which is important for multidomain setups with multiple languages).

```
ErrorException:
Warning: Attempt to read property "cssID" on null

  at vendor\contao\core-bundle\src\Controller\FrontendModule\RootPageDependentModulesController.php:46
```

The current code has a check for this in place. However, it does not actually work, because the `RootPageDependentSelect` stores a serialised array with the following data in such a case:

```
array:4 [▼
  1 => "50"
  2 => "50"
  7 => ""
  5 => ""
]
```

This PR fixes that.

Commits
-------

f637ae77 fix module if not all root pages have one defined
b7b7d293 simplify empty check, check for invalid module ID, normalize empty re…
fba9496a Merge remote-tracking branch 'origin/4.13' into fix-rootpagedependent…