Make sure login constants are set when rendering error page (see #6035)
Description
-----------

In Contao 4.13, the login constants are not set when the framework is initialized through routing, because the firewall would not yet be active. The constants are later on set by a listener that runs after the firewal. If however there is an exception between routing and the firewall, our `PrettyErrorScreenListener` will try to render the error page without the constants being set. I did get an `Warning: Use of undefined constant BE_USER_LOGGED_IN` warning in my Sentry logs.

Commits
-------

9551458f Make sure login constants are set when rendering error page
f23cb883 Set the constants through the existing listener