Fix security and CSRF issues on the command line (see #5968)
Description
-----------

Actually fixes https://github.com/codefog/contao-news_categories/issues/236

When generating content elements on the command line (in this case throught the news feed), having a `ModuleLogin` (or anything similar that uses CSRF and/or the firewall) will lead to two exceptions:
 1. the CSRF token storage will not have a token, because it is currently only initialized by a kernel request.
 2. the firewall will throw an exception because no firewall is active

The first problem can easily be solved by initializing an empty CSRF token storage on the command line. @ausi can you spot an issue here? What if we run a command from the web process, do we maybe need an `isInitialized` check first?

The second problem can be worked around because the authorization checker (which throws an error on missig firewall token) already supports using `NullToken` for the new authentication manager in Symfony 6. By simply not throwing an exception on missing tokens, we solve a lot of problems related to the firewall in my opinion. Be aware that this is also the default behaviour of Contao 5 (through Symfony 6).

Commits
-------

7e4e3518 Fix security and CSRF issues on the command line
aa897d15 Test the security.authorization_checker config
e41a3be1 Correctly test the authorization_checker config