Cache lower key values in `Model::convertToPhpValue()` (see #6438)
Description
-----------

This method is a very hot one when there are a lot of models loaded and most of the time is spent for lowering the same values over and over again. By caching them, this method becomes 5 times faster.

Commits
-------

bee684a2 Cache lower key values in Model::convertToPhpValue()
ea691711 Added comment