Add a tab menu to jump to palette sections (see #6353)
Description
-----------

Implements #5385 

The idea is to add a tab menu that allows you to jump to a section of the current palette:

<img width="1262" alt="" src="https://github.com/contao/contao/assets/1192057/e761b75f-7c00-4f73-a5c1-dd0a5962f967">

I‘m not yet happy with the design, especially if the version menu is visible as well:

<img width="1262" alt="" src="https://github.com/contao/contao/assets/1192057/f3ecd6b4-b135-4943-9824-d3425c7c2dc4">

It can also get pretty big if there are many sections:

<img width="1260" alt="" src="https://github.com/contao/contao/assets/1192057/d6e79b0c-ded0-4e64-85db-a984c576029a">

It is even worse on mobile devices:

<img width="295" alt="" src="https://github.com/contao/contao/assets/1192057/2820076e-b9d5-46c8-80b4-cb8965e719ef">

A possible solution would be to apply `overflow-x: scroll` to the parent div, so you can swipe left and right if the list exceeds the available width. We don‘t use this kind of UI anywhere else, though.

<img width="1263" alt="" src="https://github.com/contao/contao/assets/1192057/c790544f-6fca-42a2-aa87-ac154c08c3f9">

What do you think?

Commits
-------

21dff58e Revert 'Show the back end header on scroll-up (see #6569)'
0f7a2494 Add a tab menu to jump to palette sections
10419492 Add a min-height to the jump targets bar
2c7aa37a Adjust the scroll offset
079f789c Apply suggestions from code review
ea1decbc Rebuild the assets