Fix clipboard permission checks (see #6658)
Description
-----------

While working on voters in Contao 5.3, I noticed the permission checks for paste buttons are incorrect. The action currently contains the _current record_ (the one the paste button is on), but it should actually contain the record that is being copied/pasted. Depending on the clipboard new/copy/cut, the action is different and and it should contain the PID/Sorting position of the current record/paste button.

Additionally, the check on 4391/4389 was inverted (the button was disabled if access was granted).

Commits
-------

bbcf22a8 Fix clipboard permission checks
c2e615fb Fix clipboard property
da1fc6ed Only add IDs to clipboard if the user use allowed
a3172155 Set new data for specific modes
52913fc9 Re-add the $new parameter
2ede3d5d $strTable is needed to fetch records for clipboard
fdb133a6 Throw correct exception if clipboard access is denied
54f6d7c0 Empty clipboard if no IDs are allowed