Add enum support for DCAs and models (see #6584)
Description
-----------

Implements #6546 

A few notes on the implementation:

### 1. Usage of TranslatableMessage

The generation of translated references for the generated options is implemented via Symfony’s `TranslatableMessage`.

In my opinion, this is the most sane way to handle translations in a complex system with various sources of translation.

Using Symfony’s `TranslatableInterface` however is less restrictive, but does not provide the additional information on parameters and especially on the translation domain.

Although it is outside the scope of this PR, a forward-looking decision ought to be made on how Contao wants to deal with translations in the future. Up to now, neither of the two interfaces is being used.

### 2. Validation of enum values in the model

It’s not straightforward to define a proper handling of invalid enum values in the model. One could:
* simply return null for invalid values
* raise an exception explaining the invalid value (like implemented)
  * Since there is no built-in way to detect whether an enum is backed by strings or integers, only a generic exception for both types is possible (without adding somethin like an `EnumUtil`).
* let the call to `BackedEnum::tryFrom()` fail with the invalid value and let it throw its TypeError

I implemented the second way in order to give developers a proper hint pointing to the DCA configuration but the other two ways are reasonable as well.


### 3. Modifications & unit tests for legacy code

For the changes to the legacy code, I stuck to existing styles and structures as much as possible.

For the model, only a simple unit test is added for resolving the enum - if necessary, I can also add tests for the various exceptions.

### 4. Needs documentation

If this will be merged, I'll also add corresponding notes to the developer documentation.

Commits
-------

107a3591 add enum support for DCAs and models
90f9e22d generate reference independently of options
9b722df4 CS for enum cases