Implement the content URL generator (see #6596)
Description
-----------

Finally! This is a first step towards the _Content URL Generator_. I was able to get this wrapped up so that **generating the URL** is separated from **matching a URL (with parameters)**. This is only the part about generating URLs for any content, since the other part is way more complex and not finished yet.

Be aware that this deprecates the `PageModel::getFrontendUrl`/`getAbsoluteUrl` methods, but it does not replace it in all places. Probably not everywhere in core-bundle, and currently in non of the add-on bundles.

This does not yet implement content URL generator for any of our add-on bundles (like news and events), because that alone are dozens of file to adjust and will be easier to review in separate PRs (I do have some of that code available locally though, so I can quickly add them once this is merged).


- [x] Replaces https://github.com/contao/contao/pull/6547
- [x] (Probably) requires https://github.com/contao/contao/pull/6595 and https://github.com/contao/contao/pull/6594
- [x] Needs tests on the new services

Commits
-------

ba2bd8ef Implement the content URL generator
83d3f277 Various improvements, CS and tests
96368207 Use the ContentUrlGenerator in the SERP widget
ad1e1378 Correctly mark URL result as redirect
cc276765 Cache exceptions thrown when generating a content URL
279308ad Merge remote-tracking branch 'upstream/5.x' into feature/content-url-…
71c1abd1 Remove unnecessary optional parameters
53f4cdaa Re-add the reference type
bbbb67a3 Fix existing code for absolute URLs
b8177b81 Fixed tests
873e033f Skip cache if content or parameters are not serializable (e.g. contai…
1903333f Tests for ContentUrlGenerator
8fd38de8 Fixes from code review
ba10d117 Tests for resolvers
74267e8c Fixed argument of AuthenticationSuccessHandler
9a13b4e2 Fixed phpstan error
419360dc Fixed phpstan error
5487c57a Drop the obsolete SerpPreview url_callback for tl_page
03b5b566 Replace PageModel::getFrontendUrl and getAbsoluteUrl
2ea63d7c Added review comments
23b8f7b0 Fix CI
223644dd Replace PageModel::getFrontendUrl and getAbsoluteUrl
4303856b Correctly make URLs absolute
34cf9623 Fixed tests
a941ccd5 Rename the ContentUrlResult
ba24a4a3 Revert "Rename the ContentUrlResult"
9b529db7 Use NULL instead of ContentUrlResult::abstain()
f5fe0239 Remove the null option